+++ packages/ui/src/modals/AddMembershipModal.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react'
 import { ButtonPrimaryMedium } from '../components/buttons'
 import { Label, Switch, TextInput } from '../components/forms'
 import { Modal, ModalFooter, ModalHeader, ScrolledModalBody } from '../components/Modal'
-import { Help } from '../components/notifications/Help'
+import { Help } from '../components/Help'
 import { filterAccount, SelectAccount } from '../components/selects/AccountSelectTemplate/SelectAccount'
 import { TokenValue } from '../components/typography'
 import { Account } from '../hooks/types'
