+++ packages/ui/src/modals/TransferModal/SignTransferModal.tsx
@@ -7,7 +7,7 @@ import { AccountInfo } from '../../components/AccountInfo'
 import { ButtonPrimaryMedium } from '../../components/buttons'
 import { ArrowDownExpandedIcon } from '../../components/icons'
 import { Modal, ModalBody, ModalFooter, ModalHeader, SignTransferContainer } from '../../components/Modal'
-import { Help } from '../../components/notifications/Help'
+import { Help } from '../../components/Help'
 import { TokenValue } from '../../components/typography'
 import { Account } from '../../hooks/types'
 import { useApi } from '../../hooks/useApi'
