+++ packages/ui/src/pages/Profile/MyAccounts/TotalBalances.tsx
@@ -2,7 +2,7 @@ import BN from 'bn.js'
 import React from 'react'
 import styled from 'styled-components'
 import { ButtonPrimarySmall } from '../../../components/buttons'
-import { Help } from '../../../components/notifications/Help'
+import { Help } from '../../../components/Help'
 import { Label, TokenValue } from '../../../components/typography'
 import { BorderRad, Colors, Shadows } from '../../../constants'
 import { useTotalBalances } from '../../../hooks/useTotalBalances'
